service-telegram-ai-bot/requirements.txt
python-telegram-bot
httpx
requests